changed tests to have on test file with multiple test functions for rotary emb unit test.
